trace-cmd: Fix python build to handle new library headers
The python build still references the internal event-parse.h header even
if the latest libtraceevent is installed on the system. This causes
inconsistencies between what's in the updated system header with the
obsoleted internal header.

Update the make files to use the proper header to build the python code as
the rest of the build uses.

Link: https://lore.kernel.org/linux-trace-devel/20211213141601.480b173d@gandalf.local.home

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>